Update patch set 1
Patch Set 1:

From the comments near the top of the file (around line 160):  "The maximum I use is 17 minutes, which is really an empirical choice; too long and we risk silent connection loss and loss of push for that period.  Too short and we lose efficiency/battery life."

The thing is - the EAS protocol is specifically designed to "test" for periods that are too long and fall back to (and stay at) shorter periods.  Microsoft suggests a max of 30 to 45 minutes.  In my testing, most MS Exchange servers default to a max heartbeat of 30 minutes.  Exchange Server administrators are "encouraged" (by microsoft) to set the max heartbeat to something compatible with any firewall.  By extending the max to 30 minutes, there isn't a "huge" savings of battery life (as might be experienced by my other submission pending for Exchange.apk), but there is some, it lowers battery usage, and if the exchange server is servicing over 100 devices, the lowered strain on the server can be significant.

Patch-set: 1