Update patch set 1
Patch Set 1: Code-Review+1

Yes, I figured that was the case after noticing other such errors being thrown from it too. As a workaround I tried removing the lib that enabled the hardware-backed keystore to make it fallback to software-backed, which it did.

I now get anauthentication error though. I can't rule out that I have locked out the account now though but I at least verified that the password is correct.

This change of yours fixed my other previous issue though. It now seems to connect to the server so I'll +1 it for that.

Patch-set: 1
Label: Code-Review=+1