Update patch set 5
Patch Set 5:

I suspect that push notifications must be broken in the service, either because of changes here or in Email, or because of age since my office uses the newest version of EAS and Exchange, with IMAP and POP disabled, leaving only webmail and push for mobiles.

In the meantime I'll use mailcal as it just scrapes the web interface, but it would be nice to see this develop further later on, and perhaps once we know it's squared away we can port over the patches to bypass the remote wipe stuff and expose them in the UI.

Regardless, nice work, I know I get as far as logging in, it's just syncing that's broken for me.

Patch-set: 5